+++ apps/mezzanine-ui-react-hook-form-v2/CHANGELOG.md
@@ -1,3 +1,15 @@
+## 0.0.20 (2025-01-22)
+
+### 🚀 Breaking Change
+
+- rename renderFileUrl -> setFileUrl
+
+## 0.0.19 (2025-01-22)
+
+### 🚀 Breaking Change
+
+- fileHost: string -> renderFileUrl: (fileId: string) => string
+
 ## 0.0.18 (2025-01-16)
 
 ### 🚀 Features